fix: shorten CLI agent picker labels
