+++ packages/cli/src/chat/tui/forms/AgentListForm.tsx
@@ -7,6 +7,7 @@ import { Spinner } from '@inkjs/ui';
 import { computeAgentOptionsData } from '@agent/index';
 import type { AgentOptionData } from '@shared/schemas';
 import { agentName } from '@shared/schemas/agent';
+import { formatAgentOptionLabel } from '@shared/utils/agentOptionLabels';
 
 import { KeyHints } from '../ui/KeyHints';
 import { Select } from '../ui/Select';
@@ -32,6 +33,12 @@ interface AgentGroups {
 
 type AgentIdentity = Pick<AgentOptionData, 'label' | 'value'>;
 type AgentDelegationFlag = Pick<AgentOptionData, 'isOrchestrator'>;
+
+export function agentPickerLabel(
+  agent: Pick<AgentOptionData, 'label'>,
+): string {
+  return formatAgentOptionLabel(agent.label);
+}
 
 function agentDescription(agent: AgentOptionData): string {
   const source = agent.isRemote
@@ -182,7 +189,7 @@ export function AgentListForm(props: AgentListFormProps): React.JSX.Element {
   const primarySectionTitle = agentPickerPrimarySectionTitle(agents.toolUse);
   const items = agents.toolUse.map((agent) => ({
     value: agent.value,
-    label: agent.label,
+    label: agentPickerLabel(agent),
     description: agentDescription(agent),
   }));
   // The current agent may be stored as a canonical key (`source:name`) or a
@@ -195,7 +202,7 @@ export function AgentListForm(props: AgentListFormProps): React.JSX.Element {
     props.currentAgent,
   );
   const workflowRows = agents.workflow.map((agent) => ({
-    name: agent.label,
+    name: agentPickerLabel(agent),
     description: agentDescription(agent),
   }));
   const selectWindow = agentSelectWindow({
