+++ src/test-kernel/cli/AgentListForm.vitest.mts
@@ -2,6 +2,7 @@ import { describe, expect, it } from 'vitest';
 
 import {
   agentPickerPrimarySectionTitle,
+  agentPickerLabel,
   agentSelectWindow,
   currentVisibleAgent,
   hiddenCurrentAgentHint,
@@ -44,6 +45,15 @@ describe('CLI AgentListForm row budget', () => {
       currentVisibleAgent(displayAgents, 'remote:orchestrator')?.label,
     ).toBe('Orchestrator');
     expect(hiddenCurrentAgentHint(displayAgents, 'setup')).toBeUndefined();
+  });
+
+  it('trims agent picker labels before description separators', () => {
+    expect(
+      agentPickerLabel({ label: 'Engineer \u2014 software team lead' }),
+    ).toBe('Engineer');
+    expect(
+      agentPickerLabel({ label: 'Research --- derivations & numerics' }),
+    ).toBe('Research');
   });
 
   it('labels the current agent when it is hidden from the picker', () => {
